updated extras and misc
- added fun functions as a seperate module
- removed unecessary stuffs from misc
