+++ tg_bot/modules/fun_strings.py
@@ -1,0 +1,425 @@
+RUN_STRINGS = (
+    "Where do you think you're going?",
+    "Huh? what? did they get away?",
+    "ZZzzZZzz... Huh? what? oh, just them again, nevermind.",
+    "Get back here!",
+    "Not so fast...",
+    "Look out for the wall!",
+    "Don't leave me alone with them!!",
+    "You run, you die.",
+    "Jokes on you, I'm everywhere",
+    "You're gonna regret that...",
+    "You could also try /punchme, I hear that's fun.",
+    "Go bother someone else, no-one here cares.",
+    "You can run, but you can't hide.",
+    "Is that all you've got?",
+    "I'm behind you...",
+    "You've got company!",
+    "We can do this the easy way, or the hard way.",
+    "You just don't get it, do you?",
+    "Yeah, you better run!",
+    "Please, remind me how much I care?",
+    "I'd run faster if I were you.",
+    "That's definitely the droid we're looking for.",
+    "May the odds be ever in your favour.",
+    "Famous last words.",
+    "And they disappeared forever, never to be seen again.",
+    "\"Oh, look at me! I'm so cool, I can run from a bot!\" - this person",
+    "Yeah yeah, just tap /punchme already.",
+    "Here, take this ring and head to Mordor while you're at it.",
+    "Legend has it, they're still running...",
+    "Unlike Harry Potter, your parents can't protect you from me.",
+    "Fear leads to anger. Anger leads to hate. Hate leads to suffering. If you keep running in fear, you might "
+    "be the next Vader.",
+    "Multiple calculations later, I have decided my interest in your shenanigans is exactly 0.",
+    "Legend has it, they're still running.",
+    "Keep it up, not sure we want you here anyway.",
+    "You're a wiza- Oh. Wait. You're not Harry, keep moving.",
+    "NO RUNNING IN THE HALLWAYS!",
+    "Hasta la vista, baby.",
+    "Who let the dogs out?",
+    "It's funny, because no one cares.",
+    "Ah, what a waste. I liked that one.",
+    "Frankly, my dear, I don't give a damn.",
+    "My milkshake brings all the boys to yard... So run faster!",
+    "You can't HANDLE the truth!",
+    "A long time ago, in a galaxy far far away... Someone would've cared about that. Not anymore though.",
+    "Hey, look at them! They're running from the inevitable banhammer... Cute.",
+    "Han shot first. So will I.",
+    "What are you running after, a white rabbit?",
+    "As The Doctor would say... RUN!",
+)
+
+SLAP_SAITAMA_TEMPLATES = (
+    "Slap me one more time and I'll mute you.",
+    "Stop slapping me. REEEEEEEEEEEEEE.",
+    [
+        "I am muting you for a minute.", #normal reply
+        "Stop slapping me just because I can't mute you. REEEEEEEEEE.", #reply to admin
+        "tmute" #command
+    ]
+)
+
+SLAP_TEMPLATES = (
+    "{user2} was shot by {user1}.",
+    "{user2} walked into a cactus while trying to escape {user1}.",
+    "{user2} drowned whilst trying to escape {user1}.",
+    "{user2} fell into a patch of cacti.",
+    "{user2} went up in flames.",
+    "{user2} burned to death.",
+    "{user2} was burnt to a crisp whilst fighting {user1}.",
+    "{user2} was struck by lightning.",
+    "{user2} was slain by {user1}.",
+    "{user2} was killed by magic.",
+    "{user2} starved to death.",
+    "{user2} fell out of the world.",
+    "{user2} was knocked into the void by {user1}.",
+    "{user2}'s bones are scraped clean by the desolate wind.",
+    "{user2} fainted.",
+    "{user2} is out of usable Pokemon! {user2} whited out!.",
+    "{user2} is out of usable Pokemon! {user2} blacked out!.",
+    "{user2} says goodbye to this cruel world.",
+    "{user2} got rekt.",
+    "{user2} was sawn in half by {user1}.",
+    "{user2}'s melon was split by {user1}.",
+    "{user2} was sliced and diced by {user1}.",
+    "{user2}'s death put another notch in {user1}'s axe.",
+    "{user2} died from {user1}'s mysterious tropical disease.",
+    "{user2} played hot-potato with a grenade.",
+    "{user2} was knifed by {user1}.",
+    "{user2} ate a grenade.",
+    "{user2} is what's for dinner!",
+    "{user2} was terminated by {user1}.",
+    "{user2} was shot before being thrown out of a plane.",
+    "{user2} has encountered an error.",
+    "{user2} died and reincarnated as a goat.",
+    "{user1} threw {user2} off a building.",
+    "{user2} got a premature burial.",
+    "{user1} spammed {user2}'s email.",
+    "{user1} hit {user2} with a small, interstellar spaceship.",
+    "{user1} put {user2} in check-mate.",
+    "{user1} RSA-encrypted {user2} and deleted the private key.",
+    "{user1} put {user2} in the friendzone.",
+    "{user1} slaps {user2} with a DMCA takedown request!",
+    "{user2} died of hospital gangrene.",
+    "{user2} got a house call from Doctor {user1}.",
+    "{user1} beheaded {user2}.",
+    "{user2} got stoned...by an angry mob.",
+    "{user1} sued the pants off {user2}.",
+    "{user2} was one-hit KO'd by {user1}.",
+    "{user1} sent {user2} down the memory hole.",
+    "{user2} was a mistake. - '{user1}' ",
+    "{user1} checkmated {user2} in two moves.",
+    "{user2} was made redundant.",
+    "{user1} {hits} {user2} with a bat!.",
+    "{user1} {hits} {user2} with a Taijutsu Kick!.",
+    "{user1} {hits} {user2} with X-Gloves!.",    
+    "{user1} {hits} {user2} with a Jet Punch!.",
+    "{user1} {hits} {user2} with a Jet Pistol!.",
+    "{user1} {hits} {user2} with a United States of Smash!.",
+    "{user1} {hits} {user2} with a Detroit Smash!.",
+    "{user1} {hits} {user2} with a Texas Smash!.",
+    "{user1} {hits} {user2} with a California Smash!.",
+    "{user1} {hits} {user2} with a New Hampshire Smash!.",
+    "{user1} {hits} {user2} with a Missouri Smash!.",
+    "{user1} {hits} {user2} with a Carolina Smash!.",
+    "{user1} {hits} {user2} with a King Kong Gun!.",
+    "{user1} {hits} {user2} with a baseball bat - metal one.!.",
+    "*Serious punches {user2}*.",
+    "*Normal punches {user2}*.",
+    "*Consecutive Normal punches {user2}*.",
+    "*Two Handed Consecutive Normal Punches {user2}*.",
+    "*Ignores {user2} to let them die of embarassment*.",
+    "*points at {user2}* What's with this sassy... lost child?.",
+    "*Hits {user2} with a Fire Tornado*.",
+    "{user1} pokes {user2} in the eye !",
+    "{user1} pokes {user2} on the sides!",        
+    "{user1} pokes {user2}!",
+    "{user1} pokes {user2} with a needle!",
+    "{user1} pokes {user2} with a pen!",
+    "{user1} pokes {user2} with a stun gun!",
+    "{user2} is secretly a Furry!.",
+    "Hey Everybody! {user1} is asking me to be mean!",
+    "( ･_･)ﾉ⌒●~* (･.･;) <-{user2}",
+    "Take this {user2}\n(ﾉﾟДﾟ)ﾉ ))))●~* ",
+    "Here {user2} hold this\n(｀・ω・)つ ●~＊",
+    "( ・_・)ノΞ●~*  {user2},Shinaeeeee!!.",
+    "( ・∀・)ｒ鹵~<≪巛;ﾟДﾟ)ﾉ\n*Bug sprays {user2}*.",
+    "( ﾟДﾟ)ﾉ占~<巛巛巛.\n-{user2} You pest!",
+    "( う-´)づ︻╦̵̵̿╤── \(˚☐˚”)/ {user2}.",
+    "{user1} {hits} {user2} with a {item}.",
+    "{user1} {hits} {user2} in the face with a {item}.",
+    "{user1} {hits} {user2} around a bit with a {item}.",
+    "{user1} {throws} a {item} at {user2}.",
+    "{user1} grabs a {item} and {throws} it at {user2}'s face.",
+    "{user1} launches a {item} in {user2}'s general direction.",
+    "{user1} starts slapping {user2} silly with a {item}.",
+    "{user1} pins {user2} down and repeatedly {hits} them with a {item}.",
+    "{user1} grabs up a {item} and {hits} {user2} with it.",
+    "{user1} ties {user2} to a chair and {throws} a {item} at them.",
+    "{user1} gave a friendly push to help {user2} learn to swim in lava.",
+    "{user1} bullied {user2}.",
+    "Nyaan ate {user2}'s leg. *nomnomnom*",
+    "{user1} {throws} a master ball at {user2}, resistance is futile.",
+    "{user1} hits {user2} with an action beam...bbbbbb (ง・ω・)ง ====*",
+    "{user1} ara ara's {user2}.",
+    "{user1} ora ora's {user2}.",
+    "{user1} muda muda's {user2}.",
+    "{user2} was turned into a Jojo reference!",  
+    "{user1} hits {user2} with {item}.",
+)
+
+PING_STRING = (
+    "PONG!!",
+    "I am here!",
+)
+
+ITEMS = (
+    "cast iron skillet",
+    "angry meow",
+    "cricket bat",
+    "wooden cane",
+    "shovel",
+    "toaster",
+    "book",
+    "laptop",
+    "rubber chicken",
+    "spiked bat",
+    "heavy rock",
+    "chunk of dirt",
+    "ton of bricks",
+    "rasengan",
+    "spirit bomb",
+    "100-Type Guanyin Bodhisattva",
+    "rasenshuriken",
+    "Murasame",
+    "ban",
+    "chunchunmaru",
+    "Kubikiribōchō",
+    "rasengan",
+    "spherical flying kat",
+)
+
+THROW = (
+    "throws",
+    "flings",
+    "chucks",
+    "hurls",
+)
+
+HIT = (
+    "hits",
+    "whacks",
+    "slaps",
+    "smacks",
+    "bashes",
+    "pats",
+
+)
+
+ABUSE_STRINGS = (
+    "Fuck off",
+    "Stfu go fuck yourself",
+    "Ur mum gey",
+    "Ur dad lesbo",
+    "Bsdk",
+    "Nigga",
+    "Ur granny tranny",
+    "you noob",
+    "Relax your Rear,ders nothing to fear,The Rape train is finally here",
+    "Stfu bc",
+    "Stfu and Gtfo U nub",
+    "GTFO bsdk",
+    "CUnt",
+    " Gay is here",
+    "Ur dad gey bc ",
+)
+
+EYES = [
+    ['⌐■', '■'],
+    [' ͠°', ' °'],
+    ['⇀', '↼'],
+    ['´• ', ' •`'],
+    ['´', '`'],
+    ['`', '´'],
+    ['ó', 'ò'],
+    ['ò', 'ó'],
+    ['⸌', '⸍'],
+    ['>', '<'],
+    ['Ƹ̵̡', 'Ʒ'],
+    ['ᗒ', 'ᗕ'],
+    ['⟃', '⟄'],
+    ['⪧', '⪦'],
+    ['⪦', '⪧'],
+    ['⪩', '⪨'],
+    ['⪨', '⪩'],
+    ['⪰', '⪯'],
+    ['⫑', '⫒'],
+    ['⨴', '⨵'],
+    ['⩿', '⪀'],
+    ['⩾', '⩽'],
+    ['⩺', '⩹'],
+    ['⩹', '⩺'],
+    ['◥▶', '◀◤'],
+    ['◍', '◎'],
+    ['/͠-', '┐͡-\\'],
+    ['⌣', '⌣”'],
+    [' ͡⎚', ' ͡⎚'],
+    ['≋'],
+    ['૦ઁ'],
+    ['  ͯ'],
+    ['  ͌'],
+    ['ළ'],
+    ['◉'],
+    ['☉'],
+    ['・'],
+    ['▰'],
+    ['ᵔ'],
+    [' ﾟ'],
+    ['□'],
+    ['☼'],
+    ['*'],
+    ['`'],
+    ['⚆'],
+    ['⊜'],
+    ['>'],
+    ['❍'],
+    ['￣'],
+    ['─'],
+    ['✿'],
+    ['•'],
+    ['T'],
+    ['^'],
+    ['ⱺ'],
+    ['@'],
+    ['ȍ'],
+    ['  '],
+    ['  '],
+    ['x'],
+    ['-'],
+    ['$'],
+    ['Ȍ'],
+    ['ʘ'],
+    ['Ꝋ'],
+    [''],
+    ['⸟'],
+    ['๏'],
+    ['ⴲ'],
+    ['◕'],
+    ['◔'],
+    ['✧'],
+    ['■'],
+    ['♥'],
+    [' ͡°'],
+    ['¬'],
+    [' º '],
+    ['⨶'],
+    ['⨱'],
+    ['⏓'],
+    ['⏒'],
+    ['⍜'],
+    ['⍤'],
+    ['ᚖ'],
+    ['ᴗ'],
+    ['ಠ'],
+    ['σ'],
+    ['☯']
+]
+
+MOUTHS = [
+    ['v'],
+    ['ᴥ'],
+    ['ᗝ'],
+    ['Ѡ'],
+    ['ᗜ'],
+    ['Ꮂ'],
+    ['ᨓ'],
+    ['ᨎ'],
+    ['ヮ'],
+    ['╭͜ʖ╮'],
+    [' ͟ل͜'],
+    [' ͜ʖ'],
+    [' ͟ʖ'],
+    [' ʖ̯'],
+    ['ω'],
+    [' ³'],
+    [' ε '],
+    ['﹏'],
+    ['□'],
+    ['ل͜'],
+    ['‿'],
+    ['╭╮'],
+    ['‿‿'],
+    ['▾'],
+    ['‸'],
+    ['Д'],
+    ['∀'],
+    ['!'],
+    ['人'],
+    ['.'],
+    ['ロ'],
+    ['_'],
+    ['෴'],
+    ['ѽ'],
+    ['ഌ'],
+    ['⏠'],
+    ['⏏'],
+    ['⍊'],
+    ['⍘'],
+    ['ツ'],
+    ['益'],
+    ['╭∩╮'],
+    ['Ĺ̯'],
+    ['◡'],
+    [' ͜つ']
+]
+
+EARS = [
+    ['q', 'p'],
+    ['ʢ', 'ʡ'],
+    ['⸮', '?'],
+    ['ʕ', 'ʔ'],
+    ['ᖗ', 'ᖘ'],
+    ['ᕦ', 'ᕥ'],
+    ['ᕦ(', ')ᕥ'],
+    ['ᕙ(', ')ᕗ'],
+    ['ᘳ', 'ᘰ'],
+    ['ᕮ', 'ᕭ'],
+    ['ᕳ', 'ᕲ'],
+    ['(', ')'],
+    ['[', ']'],
+    ['¯\\_', '_/¯'],
+    ['୧', '୨'],
+    ['୨', '୧'],
+    ['⤜(', ')⤏'],
+    ['☞', '☞'],
+    ['ᑫ', 'ᑷ'],
+    ['ᑴ', 'ᑷ'],
+    ['ヽ(', ')ﾉ'],
+    ['\\(', ')/'],
+    ['乁(', ')ㄏ'],
+    ['└[', ']┘'],
+    ['(づ', ')づ'],
+    ['(ง', ')ง'],
+    ['⎝', '⎠'],
+    ['ლ(', 'ლ)'],
+    ['ᕕ(', ')ᕗ'],
+    ['(∩', ')⊃━☆ﾟ.*'],
+]
+
+TOSS = (
+    "Heads",
+    "Tails",
+)
+
+DECIDE = (
+    "Yes.",
+    "NoU.",
+    "Maybe."
+)
+
+TABLE = (
+    "(╯°□°）╯彡 ┻━┻",
+    "I ran out of tables, will order more.",
+    "Go do some work instead of flippin tables."
+)
