+++ tg_bot/modules/misc.py
@@ -1,0 +1,210 @@
+import html
+import json
+import re
+import requests
+
+from typing import List
+
+from telegram import Bot, Update, MessageEntity, ParseMode
+from telegram.ext import CommandHandler, run_async, Filters
+from telegram.utils.helpers import mention_html
+
+from tg_bot import dispatcher, OWNER_ID, SUDO_USERS, SUPPORT_USERS, DEV_USERS, WHITELIST_USERS
+from tg_bot.__main__ import STATS, USER_INFO, TOKEN
+from tg_bot.modules.helper_funcs.chat_status import user_admin, sudo_plus
+from tg_bot.modules.helper_funcs.extraction import extract_user
+from tg_bot.modules.disable import DisableAbleCommandHandler
+
+
+MARKDOWN_HELP = """
+Markdown is a very powerful formatting tool supported by telegram. {} has some enhancements, to make sure that \
+saved messages are correctly parsed, and to allow you to create buttons.
+
+- <code>_italic_</code>: wrapping text with '_' will produce italic text
+- <code>*bold*</code>: wrapping text with '*' will produce bold text
+- <code>`code`</code>: wrapping text with '`' will produce monospaced text, also known as 'code'
+- <code>[sometext](someURL)</code>: this will create a link - the message will just show <code>sometext</code>, \
+and tapping on it will open the page at <code>someURL</code>.
+EG: <code>[test](example.com)</code>
+
+- <code>[buttontext](buttonurl:someURL)</code>: this is a special enhancement to allow users to have telegram \
+buttons in their markdown. <code>buttontext</code> will be what is displayed on the button, and <code>someurl</code> \
+will be the url which is opened.
+EG: <code>[This is a button](buttonurl:example.com)</code>
+
+If you want multiple buttons on the same line, use :same, as such:
+<code>[one](buttonurl://example.com)
+[two](buttonurl://google.com:same)</code>
+This will create two buttons on a single line, instead of one button per line.
+
+Keep in mind that your message <b>MUST</b> contain some text other than just a button!
+""".format(dispatcher.bot.first_name)
+
+
+@run_async
+def get_id(bot: Bot, update: Update, args: List[str]):
+
+    message = update.effective_message
+    chat = update.effective_chat
+    user_id = extract_user(update.effective_message, args)
+
+    if user_id:
+
+        if update.effective_message.reply_to_message and update.effective_message.reply_to_message.forward_from:
+
+            user1 = message.reply_to_message.from_user
+            user2 = message.reply_to_message.forward_from
+
+            update.effective_message.reply_text(
+                "The original sender, {}, has an ID of <code>{}</code>.\nThe forwarder, {}, has an ID of <code>{}</code>.".format(
+                    html.escape(user2.first_name),
+                    user2.id,
+                    html.escape(user1.first_name),
+                    user1.id),
+                parse_mode=ParseMode.HTML)
+
+        else:
+
+            user = bot.get_chat(user_id)
+            update.effective_message.reply_text("{}'s id is <code>{}</code>.".format(html.escape(user.first_name), user.id),
+                                                parse_mode=ParseMode.HTML)
+
+    else:
+
+        if chat.type == "private":
+            update.effective_message.reply_text("Your id is <code>{}</code>.".format(chat.id),
+                                                parse_mode=ParseMode.HTML)
+
+        else:
+            update.effective_message.reply_text("This group's id is <code>{}</code>.".format(chat.id),
+                                                parse_mode=ParseMode.HTML)
+
+
+@run_async
+def info(bot: Bot, update: Update, args: List[str]):
+
+    message = update.effective_message
+    chat = update.effective_chat
+    user_id = extract_user(update.effective_message, args)
+
+    if user_id:
+        user = bot.get_chat(user_id)
+
+    elif not message.reply_to_message and not args:
+        user = message.from_user
+
+    elif not message.reply_to_message and (not args or (
+            len(args) >= 1 and not args[0].startswith("@") and not args[0].isdigit() and not message.parse_entities(
+        [MessageEntity.TEXT_MENTION]))):
+        message.reply_text("I can't extract a user from this.")
+        return
+
+    else:
+        return
+
+    text = "<b>Characteristics:</b>" \
+           "\nID: <code>{}</code>" \
+           "\nFirst Name: {}".format(user.id, html.escape(user.first_name))
+
+    if user.last_name:
+        text += "\nLast Name: {}".format(html.escape(user.last_name))
+
+    if user.username:
+        text += "\nUsername: @{}".format(html.escape(user.username))
+
+    text += "\nPermanent user link: {}".format(mention_html(user.id, "link"))
+
+    disaster_level_present = False
+    
+    if user.id == OWNER_ID:
+        text += "\nThe Disaster level of this person is 'God'."
+        disaster_level_present = True
+    elif user.id in DEV_USERS:
+        text += "\nThis member is one of 'Hero Association'."
+        disaster_level_present = True
+    elif user.id in SUDO_USERS:
+        text += "\nThe Disaster level of this person is 'Dragon'."
+        disaster_level_present = True
+    elif user.id in SUPPORT_USERS:
+        text+= "\nThe Disaster level of this person is 'Demon'."
+        disaster_level_present = True
+    elif user.id in WHITELIST_USERS:
+        text += "\nThe Disaster level of this person is 'Wolf'."
+        disaster_level_present = True
+
+    if disaster_level_present:
+        text += ' [<a href="https://t.me/OnePunchSupport/18340">?</a>]'
+    
+    user_member = chat.get_member(user.id)
+    if user_member.status == 'administrator':
+        result = requests.post(f"https://api.telegram.org/bot{TOKEN}/getChatMember?chat_id={chat.id}&user_id={user.id}")
+        result = result.json()["result"]
+        if "custom_title" in result.keys():
+            custom_title = result['custom_title']
+            text += f"\n\nThis user holds the title <b>{custom_title}</b> here."
+
+    for mod in USER_INFO:
+        try:
+            mod_info = mod.__user_info__(user.id).strip()
+        except TypeError:
+            mod_info = mod.__user_info__(user.id, chat.id).strip()
+        if mod_info:
+            text += "\n\n" + mod_info
+
+    update.effective_message.reply_text(text, parse_mode=ParseMode.HTML, disable_web_page_preview=True)
+
+
+@run_async
+@user_admin
+def echo(bot: Bot, update: Update):
+
+    args = update.effective_message.text.split(None, 1)
+    message = update.effective_message
+
+    if message.reply_to_message:
+        message.reply_to_message.reply_text(args[1])
+    else:
+        message.reply_text(args[1], quote=False)
+
+    message.delete()
+
+
+@run_async
+def markdown_help(bot: Bot, update: Update):
+
+    update.effective_message.reply_text(MARKDOWN_HELP, parse_mode=ParseMode.HTML)
+    update.effective_message.reply_text("Try forwarding the following message to me, and you'll see!")
+    update.effective_message.reply_text("/save test This is a markdown test. _italics_, *bold*, `code`, "
+                                        "[URL](example.com) [button](buttonurl:github.com) "
+                                        "[button2](buttonurl://google.com:same)")
+
+
+@run_async
+@sudo_plus
+def stats(bot: Bot, update: Update):
+    stats = "Current stats:\n" + "\n".join([mod.__stats__() for mod in STATS])
+    result = re.sub(r'(\d+)', r'<code>\1</code>', stats)
+    update.effective_message.reply_text(result, parse_mode=ParseMode.HTML)
+
+
+__help__ = """
+ - /id: get the current group id. If used by replying to a message, gets that user's id.
+ - /info: get information about a user.
+ - /markdownhelp: quick summary of how markdown works in telegram - can only be called in private chats.
+"""
+
+ID_HANDLER = DisableAbleCommandHandler("id", get_id, pass_args=True)
+INFO_HANDLER = DisableAbleCommandHandler("info", info, pass_args=True)
+ECHO_HANDLER = DisableAbleCommandHandler("echo", echo, filters=Filters.group)
+MD_HELP_HANDLER = CommandHandler("markdownhelp", markdown_help, filters=Filters.private)
+STATS_HANDLER = CommandHandler("stats", stats)
+
+dispatcher.add_handler(ID_HANDLER)
+dispatcher.add_handler(INFO_HANDLER)
+dispatcher.add_handler(ECHO_HANDLER)
+dispatcher.add_handler(MD_HELP_HANDLER)
+dispatcher.add_handler(STATS_HANDLER)
+
+__mod_name__ = "Misc"
+__command_list__ = ["id", "info", "echo"]
+__handlers__ = [ID_HANDLER, INFO_HANDLER, ECHO_HANDLER, MD_HELP_HANDLER, STATS_HANDLER]
