--- tg_bot/modules/misc2.py
@@ -1,498 +1,0 @@
-import html
-import json
-import os
-import psutil
-import random
-import time 
-import datetime
-from typing import Optional, List
-import re
-import requests
-from telegram import Message, Chat, Update, Bot, MessageEntity
-from telegram import ParseMode
-from telegram.ext import CommandHandler, run_async, Filters
-from telegram.utils.helpers import escape_markdown, mention_html
-from tg_bot.modules.helper_funcs.chat_status import user_admin, sudo_plus, is_user_admin
-from tg_bot import dispatcher, OWNER_ID, SUDO_USERS, SUPPORT_USERS, DEV_USERS, WHITELIST_USERS, BAN_STICKER
-from tg_bot.__main__ import STATS, USER_INFO, TOKEN
-from tg_bot.modules.disable import DisableAbleCommandHandler, DisableAbleRegexHandler
-from tg_bot.modules.helper_funcs.extraction import extract_user
-from tg_bot.modules.helper_funcs.filters import CustomFilters
-
-RUN_STRINGS = (
-    "Where do you think you're going?",
-    "Huh? what? did they get away?",
-    "ZZzzZZzz... Huh? what? oh, just them again, nevermind.",
-    "Get back here!",
-    "Not so fast...",
-    "Look out for the wall!",
-    "Don't leave me alone with them!!",
-    "You run, you die.",
-    "Jokes on you, I'm everywhere",
-    "You're gonna regret that...",
-    "You could also try /punchme, I hear that's fun.",
-    "Go bother someone else, no-one here cares.",
-    "You can run, but you can't hide.",
-    "Is that all you've got?",
-    "I'm behind you...",
-    "You've got company!",
-    "We can do this the easy way, or the hard way.",
-    "You just don't get it, do you?",
-    "Yeah, you better run!",
-    "Please, remind me how much I care?",
-    "I'd run faster if I were you.",
-    "That's definitely the droid we're looking for.",
-    "May the odds be ever in your favour.",
-    "Famous last words.",
-    "And they disappeared forever, never to be seen again.",
-    "\"Oh, look at me! I'm so cool, I can run from a bot!\" - this person",
-    "Yeah yeah, just tap /punchme already.",
-    "Here, take this ring and head to Mordor while you're at it.",
-    "Legend has it, they're still running...",
-    "Unlike Harry Potter, your parents can't protect you from me.",
-    "Fear leads to anger. Anger leads to hate. Hate leads to suffering. If you keep running in fear, you might "
-    "be the next Vader.",
-    "Multiple calculations later, I have decided my interest in your shenanigans is exactly 0.",
-    "Legend has it, they're still running.",
-    "Keep it up, not sure we want you here anyway.",
-    "You're a wiza- Oh. Wait. You're not Harry, keep moving.",
-    "NO RUNNING IN THE HALLWAYS!",
-    "Hasta la vista, baby.",
-    "Who let the dogs out?",
-    "It's funny, because no one cares.",
-    "Ah, what a waste. I liked that one.",
-    "Frankly, my dear, I don't give a damn.",
-    "My milkshake brings all the boys to yard... So run faster!",
-    "You can't HANDLE the truth!",
-    "A long time ago, in a galaxy far far away... Someone would've cared about that. Not anymore though.",
-    "Hey, look at them! They're running from the inevitable banhammer... Cute.",
-    "Han shot first. So will I.",
-    "What are you running after, a white rabbit?",
-    "As The Doctor would say... RUN!",
-)
-
-SLAP_SAITAMA_TEMPLATES = (
-    "Slap me one more time and I'll mute you.",
-    "Stop slapping me. REEEEEEEEEEEEEE.",
-    [
-        "I am muting you for a minute.", #normal reply
-        "Stop slapping me just because I can't mute you. REEEEEEEEEE.", #reply to admin
-        "tmute" #command
-    ]
-)
-
-SLAP_TEMPLATES = (
-    "{user2} was shot by {user1}.",
-    "{user2} walked into a cactus while trying to escape {user1}.",
-    "{user2} drowned whilst trying to escape {user1}.",
-    "{user2} fell into a patch of cacti.",
-    "{user2} went up in flames.",
-    "{user2} burned to death.",
-    "{user2} was burnt to a crisp whilst fighting {user1}.",
-    "{user2} was struck by lightning.",
-    "{user2} was slain by {user1}.",
-    "{user2} was killed by magic.",
-    "{user2} starved to death.",
-    "{user2} fell out of the world.",
-    "{user2} was knocked into the void by {user1}.",
-    "{user2}'s bones are scraped clean by the desolate wind.",
-    "{user2} fainted.",
-    "{user2} is out of usable Pokemon! {user2} whited out!.",
-    "{user2} is out of usable Pokemon! {user2} blacked out!.",
-    "{user2} says goodbye to this cruel world.",
-    "{user2} got rekt.",
-    "{user2} was sawn in half by {user1}.",
-    "{user2}'s melon was split by {user1}.",
-    "{user2} was sliced and diced by {user1}.",
-    "{user2}'s death put another notch in {user1}'s axe.",
-    "{user2} died from {user1}'s mysterious tropical disease.",
-    "{user2} played hot-potato with a grenade.",
-    "{user2} was knifed by {user1}.",
-    "{user2} ate a grenade.",
-    "{user2} is what's for dinner!",
-    "{user2} was terminated by {user1}.",
-    "{user2} was shot before being thrown out of a plane.",
-    "{user2} has encountered an error.",
-    "{user2} died and reincarnated as a goat.",
-    "{user1} threw {user2} off a building.",
-    "{user2} got a premature burial.",
-    "{user1} spammed {user2}'s email.",
-    "{user1} hit {user2} with a small, interstellar spaceship.",
-    "{user1} put {user2} in check-mate.",
-    "{user1} RSA-encrypted {user2} and deleted the private key.",
-    "{user1} put {user2} in the friendzone.",
-    "{user1} slaps {user2} with a DMCA takedown request!",
-    "{user2} died of hospital gangrene.",
-    "{user2} got a house call from Doctor {user1}.",
-    "{user1} beheaded {user2}.",
-    "{user2} got stoned...by an angry mob.",
-    "{user1} sued the pants off {user2}.",
-    "{user2} was one-hit KO'd by {user1}.",
-    "{user1} sent {user2} down the memory hole.",
-    "{user2} was a mistake. - '{user1}' ",
-    "{user1} checkmated {user2} in two moves.",
-    "{user2} was made redundant.",
-    "{user1} {hits} {user2} with a bat!.",
-    "{user1} {hits} {user2} with a Taijutsu Kick!.",
-    "{user1} {hits} {user2} with X-Gloves!.",    
-    "{user1} {hits} {user2} with a Jet Punch!.",
-    "{user1} {hits} {user2} with a Jet Pistol!.",
-    "{user1} {hits} {user2} with a United States of Smash!.",
-    "{user1} {hits} {user2} with a Detroit Smash!.",
-    "{user1} {hits} {user2} with a Texas Smash!.",
-    "{user1} {hits} {user2} with a California Smash!.",
-    "{user1} {hits} {user2} with a New Hampshire Smash!.",
-    "{user1} {hits} {user2} with a Missouri Smash!.",
-    "{user1} {hits} {user2} with a Carolina Smash!.",
-    "{user1} {hits} {user2} with a King Kong Gun!.",
-    "{user1} {hits} {user2} with a baseball bat - metal one.!.",
-    "*Serious punches {user2}*.",
-    "*Normal punches {user2}*.",
-    "*Consecutive Normal punches {user2}*.",
-    "*Two Handed Consecutive Normal Punches {user2}*.",
-    "*Ignores {user2} to let them die of embarassment*.",
-    "*points at {user2}* What's with this sassy... lost child?.",
-    "*Hits {user2} with a Fire Tornado*.",
-    "{user1} pokes {user2} in the eye !",
-    "{user1} pokes {user2} on the sides!",        
-    "{user1} pokes {user2}!",
-    "{user1} pokes {user2} with a needle!",
-    "{user1} pokes {user2} with a pen!",
-    "{user1} pokes {user2} with a stun gun!",
-    "{user2} is secretly a Furry!.",
-    "Hey Everybody! {user1} is asking me to be mean!",
-    "( ･_･)ﾉ⌒●~* (･.･;) <-{user2}",
-    "Take this {user2}\n(ﾉﾟДﾟ)ﾉ ))))●~* ",
-    "Here {user2} hold this\n(｀・ω・)つ ●~＊",
-    "( ・_・)ノΞ●~*  {user2},Shinaeeeee!!.",
-    "( ・∀・)ｒ鹵~<≪巛;ﾟДﾟ)ﾉ\n*Bug sprays {user2}*.",
-    "( ﾟДﾟ)ﾉ占~<巛巛巛.\n-{user2} You pest!",
-    "( う-´)づ︻╦̵̵̿╤── \(˚☐˚”)/ {user2}.",
-    "{user1} {hits} {user2} with a {item}.",
-    "{user1} {hits} {user2} in the face with a {item}.",
-    "{user1} {hits} {user2} around a bit with a {item}.",
-    "{user1} {throws} a {item} at {user2}.",
-    "{user1} grabs a {item} and {throws} it at {user2}'s face.",
-    "{user1} launches a {item} in {user2}'s general direction.",
-    "{user1} starts slapping {user2} silly with a {item}.",
-    "{user1} pins {user2} down and repeatedly {hits} them with a {item}.",
-    "{user1} grabs up a {item} and {hits} {user2} with it.",
-    "{user1} ties {user2} to a chair and {throws} a {item} at them.",
-    "{user1} gave a friendly push to help {user2} learn to swim in lava.",
-    "{user1} bullied {user2}.",
-    "Nyaan ate {user2}'s leg. *nomnomnom*",
-    "{user1} {throws} a master ball at {user2}, resistance is futile.",
-    "{user1} hits {user2} with an action beam...bbbbbb (ง・ω・)ง ====*",
-    "{user1} ara ara's {user2}.",
-    "{user1} ora ora's {user2}.",
-    "{user1} muda muda's {user2}.",
-    "{user2} was turned into a Jojo reference!",  
-    "{user1} hits {user2} with {item}.",
-)
-
-PING_STRING = (
-    "PONG!!",
-    "I am here!",
-)
-
-ITEMS = (
-    "cast iron skillet",
-    "angry meow",
-    "cricket bat",
-    "wooden cane",
-    "shovel",
-    "toaster",
-    "book",
-    "laptop",
-    "rubber chicken",
-    "spiked bat",
-    "heavy rock",
-    "chunk of dirt",
-    "ton of bricks",
-    "rasengan",
-    "spirit bomb",
-    "100-Type Guanyin Bodhisattva",
-    "rasenshuriken",
-    "Murasame",
-    "ban",
-    "chunchunmaru",
-    "Kubikiribōchō",
-    "rasengan",
-    "spherical flying kat",
-)
-
-THROW = (
-    "throws",
-    "flings",
-    "chucks",
-    "hurls",
-)
-
-HIT = (
-    "hits",
-    "whacks",
-    "slaps",
-    "smacks",
-    "bashes",
-    "pats",
-
-)
-
-
-@run_async
-def runs(bot: Bot, update: Update):
-    update.effective_message.reply_text(random.choice(RUN_STRINGS))
-
-
-@run_async
-def slap(bot: Bot, update: Update, args: List[str]):
-    msg = update.effective_message  # type: Optional[Message]
-    chat = update.effective_chat
-
-    # reply to correct message
-    reply_text = msg.reply_to_message.reply_text if msg.reply_to_message else msg.reply_text
-
-    # get user who sent message
-    if msg.from_user.username:
-        curr_user = html.escape(msg.from_user.first_name)
-    else:
-        curr_user = "{}".format(mention_html(msg.from_user.id, html.escape(msg.from_user.first_name)))
-
-    user_id = extract_user(update.effective_message, args)
-
-    if user_id == bot.id:
-        temp = random.choice(SLAP_SAITAMA_TEMPLATES)
-        if type(temp) == list:
-            if temp[2] == "tmute":
-
-                if is_user_admin(chat, msg.from_user.id):
-                    reply_text(temp[1])
-                    return
-
-                mutetime = int(time.time() + 60)
-                bot.restrict_chat_member(chat.id, msg.from_user.id, until_date=mutetime, can_send_messages=False)
-                
-            reply_text(temp[0])
-        else:
-            reply_text(temp)
-        return
-
-    if user_id:
-        slapped_user = bot.get_chat(user_id)
-        user1 = curr_user
-        if slapped_user.username:
-            user2 = html.escape(slapped_user.first_name)
-        else:
-            user2 = "{}".format(mention_html(slapped_user.id, html.escape(slapped_user.first_name)))
-
-    # if no target found, bot targets the sender
-    else:
-        user1 = "{}".format(mention_html(bot.id, bot.first_name))
-        user2 = curr_user
-
-    temp = random.choice(SLAP_TEMPLATES)
-    item = random.choice(ITEMS)
-    hit = random.choice(HIT)
-    throw = random.choice(THROW)
-
-    repl = temp.format(user1=user1, user2=user2, item=item, hits=hit, throws=throw)
-
-    reply_text(repl, parse_mode=ParseMode.HTML)
-
-
-@run_async
-def get_bot_ip(bot: Bot, update: Update):
-    """ Sends the bot's IP address, so as to be able to ssh in if necessary.
-        OWNER ONLY.
-    """
-    res = requests.get("http://ipinfo.io/ip")
-    update.message.reply_text(res.text)
-
-
-@run_async
-def get_id(bot: Bot, update: Update, args: List[str]):
-    user_id = extract_user(update.effective_message, args)
-    if user_id:
-        if update.effective_message.reply_to_message and update.effective_message.reply_to_message.forward_from:
-            user1 = update.effective_message.reply_to_message.from_user
-            user2 = update.effective_message.reply_to_message.forward_from
-            update.effective_message.reply_text(
-                "The original sender, {}, has an ID of `{}`.\nThe forwarder, {}, has an ID of `{}`.".format(
-                    escape_markdown(user2.first_name),
-                    user2.id,
-                    escape_markdown(user1.first_name),
-                    user1.id),
-                parse_mode=ParseMode.MARKDOWN)
-        else:
-            user = bot.get_chat(user_id)
-            update.effective_message.reply_text("{}'s id is `{}`.".format(escape_markdown(user.first_name), user.id),
-                                                parse_mode=ParseMode.MARKDOWN)
-    else:
-        chat = update.effective_chat  # type: Optional[Chat]
-        if chat.type == "private":
-            update.effective_message.reply_text("Your id is `{}`.".format(chat.id),
-                                                parse_mode=ParseMode.MARKDOWN)
-
-        else:
-            update.effective_message.reply_text("This group's id is `{}`.".format(chat.id),
-                                                parse_mode=ParseMode.MARKDOWN)
-
-
-@run_async
-def info(bot: Bot, update: Update, args: List[str]):
-    msg = update.effective_message  # type: Optional[Message]
-    chat = update.effective_chat # type: Optional[Chat]
-    user_id = extract_user(update.effective_message, args)
-
-    if user_id:
-        user = bot.get_chat(user_id)
-
-    elif not msg.reply_to_message and not args:
-        user = msg.from_user
-
-    elif not msg.reply_to_message and (not args or (
-            len(args) >= 1 and not args[0].startswith("@") and not args[0].isdigit() and not msg.parse_entities(
-        [MessageEntity.TEXT_MENTION]))):
-        msg.reply_text("I can't extract a user from this.")
-        return
-
-    else:
-        return
-
-    text = "<b>Characteristics:</b>" \
-           "\nID: <code>{}</code>" \
-           "\nFirst Name: {}".format(user.id, html.escape(user.first_name))
-
-    if user.last_name:
-        text += "\nLast Name: {}".format(html.escape(user.last_name))
-
-    if user.username:
-        text += "\nUsername: @{}".format(html.escape(user.username))
-
-    text += "\nPermanent user link: {}".format(mention_html(user.id, "link"))
-
-    disaster_level_present = False
-    if user.id == OWNER_ID:
-        text += "\n\nThe Disaster level of this person is 'God'."
-        disaster_level_present = True
-    elif user.id in DEV_USERS:
-        text += "\nThis member is one of 'Hero Association'."
-        disaster_level_present = True
-    elif user.id in SUDO_USERS:
-        text += "\nThe Disaster level of this person is 'Dragon'."
-        disaster_level_present = True
-    elif user.id in SUPPORT_USERS:
-        text+= "\nThe Disaster level of this person is 'Demon'."
-        disaster_level_present = True
-    elif user.id in WHITELIST_USERS:
-        text += "\nThe Disaster level of this person is 'Wolf'."
-        disaster_level_present = True
-
-    if disaster_level_present:
-        text += ' [<a href="https://t.me/OnePunchSupport/18340">?</a>]'
-    
-    user_member = chat.get_member(user.id)
-    if user_member.status == 'administrator':
-        result = requests.post(f"https://api.telegram.org/bot{TOKEN}/getChatMember?chat_id={chat.id}&user_id={user.id}")
-        result = result.json()["result"]
-        if "custom_title" in result.keys():
-            custom_title = result['custom_title']
-            text += f"\n\nThis user holds the title <b>{custom_title}</b> here."
-
-    for mod in USER_INFO:
-        try:
-            mod_info = mod.__user_info__(user.id).strip()
-        except TypeError:
-            mod_info = mod.__user_info__(user.id, chat.id).strip()
-        if mod_info:
-            text += "\n\n" + mod_info
-
-    update.effective_message.reply_text(text, parse_mode=ParseMode.HTML, disable_web_page_preview=True)
-
-
-@run_async
-@user_admin
-def echo(bot: Bot, update: Update):
-    args = update.effective_message.text.split(None, 1)
-    message = update.effective_message
-    if message.reply_to_message:
-        message.reply_to_message.reply_text(args[1])
-    else:
-        message.reply_text(args[1], quote=False)
-    message.delete()
-
-
-MARKDOWN_HELP = """
-Markdown is a very powerful formatting tool supported by telegram. {} has some enhancements, to make sure that \
-saved messages are correctly parsed, and to allow you to create buttons.
-
-- <code>_italic_</code>: wrapping text with '_' will produce italic text
-- <code>*bold*</code>: wrapping text with '*' will produce bold text
-- <code>`code`</code>: wrapping text with '`' will produce monospaced text, also known as 'code'
-- <code>[sometext](someURL)</code>: this will create a link - the message will just show <code>sometext</code>, \
-and tapping on it will open the page at <code>someURL</code>.
-EG: <code>[test](example.com)</code>
-
-- <code>[buttontext](buttonurl:someURL)</code>: this is a special enhancement to allow users to have telegram \
-buttons in their markdown. <code>buttontext</code> will be what is displayed on the button, and <code>someurl</code> \
-will be the url which is opened.
-EG: <code>[This is a button](buttonurl:example.com)</code>
-
-If you want multiple buttons on the same line, use :same, as such:
-<code>[one](buttonurl://example.com)
-[two](buttonurl://google.com:same)</code>
-This will create two buttons on a single line, instead of one button per line.
-
-Keep in mind that your message <b>MUST</b> contain some text other than just a button!
-""".format(dispatcher.bot.first_name)
-
-
-@run_async
-def markdown_help(bot: Bot, update: Update):
-    update.effective_message.reply_text(MARKDOWN_HELP, parse_mode=ParseMode.HTML)
-    update.effective_message.reply_text("Try forwarding the following message to me, and you'll see!")
-    update.effective_message.reply_text("/save test This is a markdown test. _italics_, *bold*, `code`, "
-                                        "[URL](example.com) [button](buttonurl:github.com) "
-                                        "[button2](buttonurl://google.com:same)")
-
-
-@run_async
-def stats(bot: Bot, update: Update):
-    stats = "Current stats:\n" + "\n".join([mod.__stats__() for mod in STATS])
-    result = re.sub(r'(\d+)', r'<code>\1</code>', stats)
-    update.effective_message.reply_text(result, parse_mode=ParseMode.HTML)
-
-
-# /ip is for private use
-__help__ = """
- - /id: get the current group id. If used by replying to a message, gets that user's id.
- - /runs: reply a random string from an array of replies.
- - /slap: slap a user, or get slapped if not a reply.
- - /time <place>: gives the local time at the given place.
- - /info: get information about a user.
-
- - /markdownhelp: quick summary of how markdown works in telegram - can only be called in private chats.
-"""
-
-__mod_name__ = "Misc"
-
-ID_HANDLER = DisableAbleCommandHandler("id", get_id, pass_args=True)
-IP_HANDLER = CommandHandler("ip", get_bot_ip, filters=Filters.chat(SUDO_USERS + DEV_USERS))
-
-RUNS_HANDLER = DisableAbleCommandHandler("runs", runs)
-SLAP_HANDLER = DisableAbleCommandHandler("slap", slap, pass_args=True)
-INFO_HANDLER = DisableAbleCommandHandler("info", info, pass_args=True)
-SLAP_REGEX_HANDLER = DisableAbleRegexHandler("(?i)bhag", slap, friendly="slap")
-ECHO_HANDLER = CommandHandler("echo", echo, filters=Filters.group)
-MD_HELP_HANDLER = CommandHandler("markdownhelp", markdown_help, filters=Filters.private)
-STATS_HANDLER = CommandHandler("stats", stats, filters=CustomFilters.sudo_filter | CustomFilters.dev_filter)
-
-dispatcher.add_handler(ID_HANDLER)
-dispatcher.add_handler(IP_HANDLER)
-dispatcher.add_handler(RUNS_HANDLER)
-dispatcher.add_handler(SLAP_HANDLER)
-dispatcher.add_handler(INFO_HANDLER)
-dispatcher.add_handler(ECHO_HANDLER)
-dispatcher.add_handler(MD_HELP_HANDLER)
-dispatcher.add_handler(STATS_HANDLER)
-dispatcher.add_handler(SLAP_REGEX_HANDLER)
